+++ libraries/utilities/src/IndentedTextWriter.cpp
@@ -126,9 +126,9 @@ namespace utilities
         }
     }
 
-    void IndentedTextWriter::WriteIndent(int count)
+    void IndentedTextWriter::WriteIndent(size_t count)
     {
-        for (int i = 0; i < count; ++i)
+        for (size_t i = 0; i < count; ++i)
         {
             _buffer << _tabString;
         }
